Make linux_sendsig and linux_sigreturn use all 64 bits of a linux_sigset_t by updating the linux_sigframe struct so as to include linux's "extramask" field.  This field contains the upper 32-bits of the sigset.  extramask sits behind a linux_fpstate struct, which I've defined primarily for padding purposes.
While we're here, define LINUX_NSIG in terms of LINUX_NBPW (32) and
LINUX_NSIG_WORDS (2).

This fixes problems where threaded apps would accumulate a large
number of zombies.  This was happening because the exit signal resides
in the upper 32-bits of the sigset and was never getting unmasked by
the manager thread after the first child exited.

PR: 		misc/18530  (may be related, originator not yet contacted)
Reviewed by:	marcel


git-svn-id: svn+ssh://svn.freebsd.org/base/head@68689 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f