Hold the currently selected vnode's lock across the call to VOP_GETVOBJECT.
Don't try to create a vm object before the file system has a chance to finish
initializing it.  This is incorrect for a number of reasons.  Firstly, that
VOP requires a lock which the file system may not have initialized yet. Also,
open and others will create a vm object if it is necessary later.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@96094 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f