After some comments from bde, rewrite the loops to avoid turning the previously used "micro-optimization" (count-down loop) into a pessimization.  Now the loops are written in the more natural count-up form.
Also, while being there, i made the logic in out_fdc() similar to the
logic in in_fdc().  The old implementation was a bit bogus anyway
since it first tested the DIO bit and only afterwards the RQM bit.
However, according to the description of the i82077, the DIO bit is
only guaranteed to be valid once the RQM bit is set.  Thus, the old
implementatoin would have had the chance to misbehave on a controller
that is implemented in accordance with the i82077 description (but is
not bug-for-bug compatible).

MFC after:	3 days


git-svn-id: svn+ssh://svn.freebsd.org/base/head@96611 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f