Groff 1.18.1 comes in with ANSI color support, enabled by default. In "nroff" mode, italic font renders as an underlined text, which makes it indistinguishable from the bold text on color monitors (cons25 terminal type), yet it requires the less(1)'s -R option. (Refer to the new grotty(1) manpage for details.)
So turn off the color support for now (when generating catpages),
until we figure out what do we do with this new feature.  I have
a patch for grotty(1) that tells it to use the "reverse video"
attribute to render the italic font.  Once this is accepted, we
can turn color support back on (if there won't be any objections
from the community).


git-svn-id: svn+ssh://svn.freebsd.org/base/head@104874 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f