Make changes such that the kernel will now fit on floppy #1.
This involves moving all networking drivers to floppy 2, all RAID
nfsclient && cd9660 to floppy 3 (and adding ispfw).

COMPAT_FREEBSD4 had to come back, maxusers is down to 2.

AHC is out until somebody can make it a module for alpha.

TurboLaser is out because it rarely has a floppy and probably
no longer works *anyway*.

There's probably some cleanup that can be done because I was told
that things in drivers.conf don't have to be removed via dokern.sh,
but because I formally tested the release build with this version
and don't want to spend another half day on it, it's going in.

I'm sure we'll have to do some further changes prior to release.

NB: Not tested end to end yet because I've been told that sysinstall
doesn't yet work.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@106620 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f