The use of ld(1) to strip compiler local and non-global symbols from object files has bitrotted over the last thirteen years, and it now does more harm than good.
An attempt to work around the problems caused by using
ld(1) for stripping was to pass LDFLAGS to the ld(1)
command, but this was not right either as ${LDFLAGS}
should, by design, be used with cc(1) and not ld(1).

One of the proposed solutions was to use the objcopy(1)
utility to do the strip work, and the other would be to
use strip(1), but Bruce Evans suggested not stripping
any symbols at all.  This works by leaving the grunt
work to the final strip(1) command (when installing the
binary).

Submitted by:	bde


git-svn-id: svn+ssh://svn.freebsd.org/base/head@117080 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f